int in Lapack and Blas wrapper is changed to intptr_t. ILP64 interface of Lapack and Blas (MKL and OpenBLAS) is used. 32-bits machines are not tested yet. #30161

+++ src/base/MathUtilities.cpp
@@ -142,8 +142,8 @@ void CalcRotatingMatrix(double matrix[][3], double sita, CartesianType cartesian
 // calculate determinant of the matrix
 double GetDeterminant(double** matrix, int dim){
    double determinant=1.0;
-   int* ipiv=NULL;
-   MallocerFreer::GetInstance()->Malloc<int>(&ipiv, dim);
+   intptr_t* ipiv=NULL;
+   MallocerFreer::GetInstance()->Malloc<intptr_t>(&ipiv, dim);
    MolDS_wrappers::Lapack::GetInstance()->Dgetrf(matrix, ipiv, dim, dim);
    for(int i=0; i<dim; i++){
       determinant*=matrix[i][i];
@@ -151,7 +151,7 @@ double GetDeterminant(double** matrix, int dim){
          determinant *= -1.0;
       }
    }
-   MallocerFreer::GetInstance()->Free<int>(&ipiv, dim);
+   MallocerFreer::GetInstance()->Free<intptr_t>(&ipiv, dim);
    return determinant;
 }
 }
